Update `WorkGiver_HaulRottable.cs`
The same issues exists here too. NOTE: I do not understand why the other function checks the same members for `null` but not this. I left this code as is, it's up to you to decide if you wish to do something about that. I'd suggest to verify if they could plausibly be `null` and remove the checks.